Adapters cleaned
