+++ app/src/main/res/layout-sw600dp/activity_tabbed.xml
@@ -48,12 +48,4 @@
 
     </LinearLayout>
 
-    <FrameLayout
-        android:id="@+id/full_info_cont"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"
-        android:layout_margin="100dp"/>
-
 </RelativeLayout>
